requirements.txt
numpy>=1.21.0
matplotlib>=3.4.0
seaborn>=0.11.0
reportlab>=3.6.8
Pillow>=8.3.0
python-mnist>=0.7 